+++ src/NeuralNet/Layers/PReLU.php
@@ -4,8 +4,8 @@ namespace Rubix\ML\NeuralNet\Layers;
 
 use Rubix\Tensor\Matrix;
 use Rubix\ML\NeuralNet\Deferred;
-use Rubix\ML\NeuralNet\Parameter;
-use Rubix\ML\NeuralNet\VectorParam;
+use Rubix\ML\NeuralNet\Parameters\Parameter;
+use Rubix\ML\NeuralNet\Parameters\VectorParam;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
 use Rubix\ML\NeuralNet\Initializers\Constant;
 use Rubix\ML\NeuralNet\Initializers\Initializer;
@@ -46,7 +46,7 @@ class PReLU implements Hidden, Parametric
     /**
      * The parameterized leakage coeficients.
      *
-     * @var \Rubix\ML\NeuralNet\Parameter|null
+     * @var \Rubix\ML\NeuralNet\Parameters\Parameter|null
      */
     protected $alpha;