+++ src/NeuralNet/Optimizers/StepDecay.php
@@ -3,7 +3,7 @@
 namespace Rubix\ML\NeuralNet\Optimizers;
 
 use Rubix\Tensor\Tensor;
-use Rubix\ML\NeuralNet\Parameter;
+use Rubix\ML\NeuralNet\Parameters\Parameter;
 use InvalidArgumentException;
 
 /**
@@ -79,7 +79,7 @@ class StepDecay implements Optimizer
     /**
      * Take a step of gradient descent for a given parameter.
      *
-     * @param \Rubix\ML\NeuralNet\Parameter $param
+     * @param \Rubix\ML\NeuralNet\Parameters\Parameter $param
      * @param \Rubix\Tensor\Tensor $gradient
      */
     public function step(Parameter $param, Tensor $gradient) : void